nfc: move resource range analysis to `RootSignatureValidations` library
