+++ llvm/unittests/Frontend/HLSLRootSignatureRangesTest.cpp
@@ -6,7 +6,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "llvm/Frontend/HLSL/HLSLRootSignature.h"
+#include "llvm/Frontend/HLSL/RootSignatureValidations.h"
 #include "gtest/gtest.h"
 
 using namespace llvm::hlsl::rootsig;
